requirements.txt
langchain
langchain_openai
python-dotenv
langchain_community
langchain_core
streamlit
langchain_groq
faiss-cpu
langchain-text-splitters
pypdf
pymupdf
sentence_transformers
langchain-ollama
chromadb
langchain_huggingface
langchain-chroma
arxiv
wikipedia
